tests/requirements-test.txt
# Testing dependencies for VibeCortex
pytest>=7.0.0
pytest-cov>=4.0.0
Pillow>=9.0.0
